Milestone reached for PhotoGlass... Time for sleep.
Loads some pictures and put them in a wall formation.
Arrow keys navigate highlighting
BUGS: 
The QLL has some quirky linkage, will crash xor by pressing right
in the top right of the 5x5.

Up is down and down is up?! Is this an XOR issue?

The first picture doesn't start highlighted.

From here:
Should be easy to add new image loaders. Think multicast and XML.
After XML, PhotoGlass will have all of the functionality of 
PhotoAlbum and more.  
Once a MulticastImageLoader is written and a way to trigger
the creation of one is devised it will only be a matter of
writing a daemon to get live photo updating working.