This is a pretty big commit. I added a new MulticastSocket.  The Data socket.
It runs on port 1239 and also provides implements for reading
bytes from the socket.

Added RWops ability to Texture and TextureFactory.

PhotoGlass got a MulticastPictureLoader.  Use the pglassdaemon
in the lapyx folder to send some sample pictures.  It seems
horribly buggy.  I haven't a clue what the issue is, but at
least it's a proof of concept.